Fix compile warnings (#4)
* Fix compile warnings

unused-variable @ allegro.h:612, allegro.h:583
  fixed with [[maybe_unused]] attribute

misleading-indentation @ allegro.h:685
  fixed by adding curly brackets around if statement body to remove any
  ambiguity

C4244 conversion from 'double' to 'int' @ allegro.cpp:3454, allegro.cpp:3469
  fixed with explicit static_cast<int>()

C4244 conversion from 'int64' to 'double' @ allegrord.cpp:634
  fixed with explicit static_cast<double>()

* (void) cast instead of [[maybe_unused]]

* Use C-style casts

* Formatting fix

* Fix if statement scope

* Restore while loops

+++ allegro.cpp
@@ -3451,7 +3451,7 @@ void Alg_seq::beat_to_measure(double beat, int *measure, double *m_beat,
             prev_den = time_sig[tsx].den;
         } else {
             m = m + (beat - prev_beat) / bpm;
-            *measure = m;
+            *measure = (int)(m);
             *m_beat = (m - *measure) * bpm;
             *num = prev_num;
             *den = prev_den;
@@ -3466,7 +3466,7 @@ void Alg_seq::beat_to_measure(double beat, int *measure, double *m_beat,
     }
     bpm = prev.num * 4 / prev.den;
     m = m + (beat - prev.beat) / bpm;
-    *measure = m;
+    *measure = (int)(m);
     *m_beat = (m - *measure) * bpm;
     *num = prev.num;
     *den = prev.den;
